# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: SensiLog API
  version: 1.0.0
  description: VALORANT感度・設定管理API
  contact:
    name: SensiLog Team
    url: https://sensilog.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sensilog.com
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /settings/records:
    get:
      summary: 設定記録一覧取得
      tags: [Settings]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: 取得件数
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: オフセット
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: タグフィルター
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: 開始日
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: 終了日
      responses:
        200:
          description: 設定記録一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/SettingsRecord'
                  total:
                    type: integer
                    description: 総件数
                  hasMore:
                    type: boolean
                    description: 次ページ有無
        401:
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: 設定記録作成
      tags: [Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettingsRecord'
      responses:
        201:
          description: 作成された設定記録
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsRecord'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
  /settings/records/{id}:
    get:
      summary: 設定記録詳細取得
      tags: [Settings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 設定記録詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsRecord'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: 設定記録更新
      tags: [Settings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRecord'
      responses:
        200:
          description: 更新された設定記録
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsRecord'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: 設定記録削除
      tags: [Settings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: 削除成功
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /settings/suggestions:
    get:
      summary: 設定入力サジェスト取得
      tags: [Settings]
      security:
        - bearerAuth: []
      responses:
        200:
          description: サジェスト一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  mice:
                    type: array
                    items:
                      type: string
                    description: マウスデバイス一覧
                  keyboards:
                    type: array
                    items:
                      type: string
                    description: キーボード一覧
                  mousepads:
                    type: array
                    items:
                      type: string
                    description: マウスパッド一覧
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
...
components:
  schemas:
...
    SettingsRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        sensitivity:
          type: number
          format: float
          minimum: 0
          maximum: 10
          description: 感度
        dpi:
          type: integer
          minimum: 100
          maximum: 10000
          description: DPI
        mouseDevice:
          type: string
          maxLength: 100
          description: マウスデバイス
        keyboardDevice:
          type: string
          maxLength: 100
          description: キーボードデバイス
        mousepad:
          type: string
          maxLength: 100
          description: マウスパッド
        tags:
          type: array
          items:
            type: string
          description: タグ
        comment:
          type: string
          maxLength: 500
          description: コメント
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - sensitivity
        - dpi
        - createdAt
    CreateSettingsRecord:
      type: object
      properties:
        sensitivity:
          type: number
          format: float
          minimum: 0
          maximum: 10
        dpi:
          type: integer
          minimum: 100
          maximum: 10000
        mouseDevice:
          type: string
          maxLength: 100
        keyboardDevice:
          type: string
          maxLength: 100
        mousepad:
          type: string
          maxLength: 100
        tags:
          type: array
          items:
            type: string
        comment:
          type: string
          maxLength: 500
      required:
        - sensitivity
        - dpi
    UpdateSettingsRecord:
      type: object
      properties:
        sensitivity:
          type: number
          format: float
          minimum: 0
          maximum: 10
        dpi:
          type: integer
          minimum: 100
          maximum: 10000
        mouseDevice:
          type: string
          maxLength: 100
        keyboardDevice:
          type: string
          maxLength: 100
        mousepad:
          type: string
          maxLength: 100
        tags:
          type: array
          items:
            type: string
        comment:
          type: string
          maxLength: 500
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Hex色コード
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        code:
          type: string
          description: エラーコード
        details:
          type: object
          description: 詳細情報
      required:
        - error
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT